Readable language labels + 0.5°C target steps (v0.4.2)
- Display language select now uses readable, localized option labels
  (English/Engels/Englisch/Anglais, ...) instead of raw codes via slug
  options + state translations in en/nl/de/fr.
- Climate target temperature now steps in 0.5 °C when the display is
  Celsius (1 °F when Fahrenheit). Verified live that the CLI stores the
  target only in whole °F, so a 0.5 °C request maps to the nearest whole
  °F (~0.56 °C resolution) — finer control than the old 1 °C step.

Also verified live that the ketl= 'ipb' flag does NOT track lift state
(stays 0 on base and lifted), so lift detection keeps using tempr=nan.

+++ custom_components/fellow_stagg/kettle_http.py
@@ -166,7 +166,10 @@ class KettleHttpClient:
     state = "S_Heat" if power_on else "S_Off"
     await self._cli_command(session, f"setstate {state}")
 
-  async def async_set_temperature(self, session: ClientSession, temp_c: int, **_: Any) -> None:
+  async def async_set_temperature(self, session: ClientSession, temp_c: float, **_: Any) -> None:
+    # The CLI stores the target as whole degrees Fahrenheit (verified live: it does
+    # not accept 0.5 °C or fractional values). Round to the nearest whole °F so a
+    # 0.5 °C request lands on the closest achievable value (~0.56 °C resolution).
     temp_f = round((temp_c * 1.8) + 32.0)
     await self._cli_command(session, f"setsetting settempr {temp_f}")
